Use variable substitution for certain OVS paths.
The eventual goal is to not require ovs to be included as a git
sutree. This is an easy first step towards doing so. This allows for
variables to be defined at configure time that specify locations for OVS
files.

This is not comprehensive, because it does not create variables for
things like OVS utilities (e.g. ovs-pcap) or OVS test programs (e.g.
ovs-test). However, it lays the foundation for further improvement.

The variables added in this commit are:

* OVS_SRCDIR: The top-level OVS source directory.
* OVS_BUILDDIR: The top-level OVS build directory.
* OVS_MANDIR: The directory containing manpages included by OVN.
* OVS_LIBDIR: The directory containing libovs.la
* OVSDB_LIBDIR: The directory containing libovsdb.la

By default, these variables are defined as if ovs exists as a subtree in
the ovn repo.