Update the README and add a TODO_SPLIT document.
The README is based on the OVS README and likely could use some
improvement from someone a little better at marketing software than me.

The TODO_SPLIT document lists some known tasks that need to be done in
order to complete the split of OVN from OVS. This doc focuses mainly on
the technical aspects of the split rather than the social,
administrative, or policy angles of the split.