Handle GARP reply packets from provider networks only on gateway chassis
Suppose there is a provider network (with localnet port) and it is
connected to a logical router via a distributed gateway router port.
When an external switch sends GARP request packets, these packets
enter the br-int (via the patch port from the provider bridge) and
the action - put_arp is applied in the router pipeline only on the
gateway chassis where the distributed gateway router port is scheduled.
Other chassis even if they receive don't apply this action. This functionality
was added by the commit [1].

But that is not the case with GARP reply packets sent by the external
switch. ovn-controllers running on all the chassis configured with
ovn-bridge-mappings receive these packets as packet-ins because
of put_arp action. This results in unnecessary processing of these
packets only to be ignored. pinctrl thread wakes up the main ovn-controller
thread wasting few CPU cycles. It is enough for the gateway chassis to
handle these packets where the distributed router gateway port is scheduled.

This patch achieves the same - similar to GARP request packets. The below
logical flows are added
 - table=1 (lr_in_ip_input), priority=92,
   match=(inport == "lrp" &&
          !is_chassis_resident("cr-lrp") && eth.bcast &&
          arp.op == 2 && arp.spa == PROVIDER_NETWORK_CIDR), action=(drop;)

 - table=1 (lr_in_ip_input), priority=92
   match=(inport == "lr0-public" && is_chassis_resident("cr-lr0-public")
          && eth.bcast && arp.op == 2 && arp.spa == PROVIDER_NETWORK_CIDR),
   action=(put_arp(inport, arp.spa, arp.sha);)

This patch doesn't affect the arp replies from the overlay networks in the
router pipeline. This only handles GARP replies from the provider networks.

In the present master this is not much of any issue even if ovn-controller
wakes up, thanks to incremental processing engine. But in the older
versions - 2.11, 2.10 and 2.9, this results in complete flow calculations
resulting in much more CPU cyles. This patch will definitely help in saving
these CPU cyles if backported.

[1] - 3dd9febe953f("ovn-northd: Support learning neighbor from ARP request.")

CC: Han ZHou <hzhou8@ebay.com>
Acked-by: Han ZHou <hzhou8@ebay.com>
Signed-off-by: Numan Siddique <nusiddiq@redhat.com>