ovn-openstack.rst: Add check for Gateway_Chassis table
This is related to the section called "Adding a Gateway".

Added workaround command for a known issue in networking-ovn.
Even after the issue is resolved, it may be useful to have this
in the tutorial, so folks have a feel for how OVN keeps track
of chassis gateway.

Also, removed redundant route command when assigning an address
to the br-ex interface.

Signed-off-by: Flavio Fernandes <flavio@flaviof.com>