Update project name in RST file headers.
All of the RST files referred to "Open vSwitch documentation".  Do a
big search and replace to change this to "OVN documentation" since OVN
is now in its own repository.

I started doing this by hand in a handful of files as I modified them.
I figured I'd just do an automated bulk pass at it instead.  I realize
that many of the files touched may not be relevant anymore, but the
change seems harmless enough.

Signed-off-by: Russell Bryant <russell@ovn.org>
Acked-by: Ben Pfaff <blp@ovn.org>