Acknowledge that OVN committers are a new group.
The MAINTAINRES.rst and related files still discussed the OVS
committers group.  Since OVN was split out into its own repository,
the group of people with commit rights to OVN is different, and has
evolved to include a couple of people that were not previously OVS
committers.

This change aims to further formalize the creation of a separate group
of OVN committers.  It includes the OVS committers that made the most
changes to OVN in the past, as well as those who have already been
granted commit rights to OVN.

This change proposes that the new OVN committers group continues to
follow the exact same policies and procedures as the OVS committers
group, and simply updates the project name in those documents.

Signed-off-by: Russell Bryant <russell@ovn.org>
Acked-by: Ben Pfaff <blp@ovn.org>