configure: Improve checks for OVS source and build directories.
Until now, "configure" was willing to accept anything as the OVS source
and build directory, but obviously only an actual OVS source or build
directory will actually work.  This commit adds some basic checking.

In addition, this changes relative paths so that they are relative to
the current working directory when "configure" is invoked, instead of
relative to the location of the "configure" script.  This is less
surprising.

Finally, this moves the `eval echo "$dir"` lines up higher so that they
do their intended job of expanding ~ in the right place.  That has to
happen early, otherwise, ~/foo will effectively end up as $PWD/\~/foo.

Acked-by: Numan Siddique <nusiddiq@redhat.com>
Signed-off-by: Ben Pfaff <blp@ovn.org>