ovn-lib.in: Fix ovn-ctl status_xxx.
Currently ovn-ctl status_xxx (e.g. status_northd) shows "not running" even
when the daemon is running, because it is calling daemon_status() which is
defined in ovs-ctl using OVS rundir instead of OVN rundir.

A new function ovn_daemon_status was introduced to replace daemon_status
as part of ovn/ovs split.  We could fix the problem by replacing all
daemon_status() call to ovn_daemon_status() call.  However, this is not
necessary because when same function name is defined in both ovn-lib
and ovs-lib, the one in ovn-lib will take effect.  So this patch simply
renames the ovn_daemon_status() back to daemon_status().

Acked-by: Numan Siddique <numans@ovn.org>
Signed-off-by: Han Zhou <hzhou@ovn.org>