Send service monitor health checks
ovn-controller will periodically sends out the service monitor packets
for the services configured in the SB DB Service_Monitor table. This
patch makes use of the action - handle_svc_check to handle the service
monitor reply packets from the service.

This patch supports IPv4 TCP and UDP service monitoring. For TCP services,
it sends out a TCP SYN packet and expects TCP ACK packet in response.
If the response is received on time, the status of the service is set to
"online", otherwise it is set to "offline".

For UDP services, it sends out a empty UDP packet and doesn't expect any
reply. In case the service is down, the host running the service, sends out
ICMP type 3 code 4 (destination unreachable) packet. If ovn-controller receives this
ICMP packet, it sets the status of the service to "offline".

Right now only IPv4 service monitoring is supported. An upcoming patch will add
the support for IPv6.

Acked-by: Mark Michelson <mmichels@redhat.com>
Signed-off-by: Numan Siddique <numans@ovn.org>