Revert conjunctive match removal patches.
This partially reverts commits 6f914327a55aaab11507db0911b08d695e17ce2a
and 298701dbc99645700be41680a43d049cb061847a. This restores the behavior
of making conjunctive matches, and it restores the tests to their state.

The one thing it does not revert is the addition of the
force_cross_product variable in expr.c. This variable will be used in a
later commit in this series. Instead of removing it, we just set it
"false" to restore previous behavior.

With this commit, the conjunctive match misbehavior described in commit
298701dbc99645700be41680a43d049cb061847a is restored. However, this will
be fixed in an upcoming commit in this series.

Signed-off-by: Mark Michelson <mmichels@redhat.com>
Acked-by: Numan Siddique <nusiddiq@redhat.com>