controller: Downgrade a warning log message
This log message was introduced in commit 5344f24ecb.  It gets hit
under normal circumstances, so it would be better as a debug message
instead of a warning.  I also expanded it to clarify that the next
step will be to create the chassis record.

This was found by trying to run the system-ovn.at tests, and they
failed because of these unexpected warning log messages.

Signed-off-by: Russell Bryant <russell@ovn.org>
Acked-by: Numan Siddique <numans@ovn.org>