northd/automake.mk: Extract deps from .la files for cargo.
ovn-northd-ddlog depends on `libopenvswitch.a` and `libovn.a`.
Therefore, when we compile a standalone executable (enabled by the
`--enable-ddlog-northd-cli` switch to `configure`), it must be linked
against these libraries *and* their dependencies.  The latter is tricky,
since `cargo` (the Rust build system) does not understand libtool files,
so we don't have an easy way to pull in dependencies automatically.
Until now, we dealt with this by manually listing dependencies in the
`cargo` command line.  This no longer works as OVS now has at least one
configurable dependency, the `libunwind.a` library, that is enabled by
the OVS configuration script.  This results is link errors with the latest
version of OVS.

The only solution I could come up with is to extract dependencies from
`libopenvswitch.a` and `libovn.a` using `grep` and pass them to cargo.

Signed-off-by: Leonid Ryzhyk <ryzhyk@ovn.org>