nbctl: Log the source of duplicate IP addresses
When doing an lsp-add, ovn-nbctl will ensure that there is not another
port on the same logical switch with the same IP address.  I'm seeing
this error occur with ovn-kubernetes, and I would find it helpful to
see which port it thinks had the duplicate address, because it's not
obvious what's happening.

Signed-off-by: Russell Bryant <russell@ovn.org>
Acked-by: Han Zhou <hzhou@ovn.org>