tests: reset pcap before comparing with expected RA
Unless we reset the pcap before extracting the received RA frame from
it, there is a chance that a previous test left a frame already
captured in the file, in which case we'll compare the expected frame
with a frame that does not belong to the current test check.
Alternatively, an old frame will be left in the pcap file AND the new
frame will be appended to it, resulting in two frames captured instead
of one.

Both scenarios result in spurious failure of the RA tests.

The fix is to reset the captured traffic file at the start of each
check for each hypervisor.

Signed-off-by: Ihar Hrachyshka <ihrachys@redhat.com>
Signed-off-by: Ben Pfaff <blp@ovn.org>