ovn-controller: Avoid creating patch port for non-local datapaths.
When external-ids:ovn-monitor-all is set to true, patch ports for
non-local datapaths may be created, which is unnecessary and
confusing.  This patch avoids that by checking if a localnet port
belongs to local datapaths before creating the patch port.  It
also moves patch_run() in mainloop after engine_run(), so that
local_datapaths already reflects the up-to-date situation when
patch_run() is called.

Reported-by: Girish Moodalbail <gmoodalbail@nvidia.com>
Acked-by: Dumitru Ceara <dceara@redhat.com>
Signed-off-by: Han Zhou <hzhou@ovn.org>