northd: Allow /64 after ipv6_prefix
We recently hit a bug in ovn-kubernetes, where I accidentally added
/64 at the end of ipv6_prefix, to match the format we used for the
subnet option for IPv4.  This was not allowed.

This patch update ovn-northd to take the ipv6_prefix either with or
without a trailing "/64".  It still enforces a /64 CIDR prefix length.

A test case was updated to ensure that a prefix with "/64" is now
accepted.

Signed-off-by: Russell Bryant <russell@ovn.org>
Acked-by: Numan Siddique <numans@ovn.org>