Remove support for obsolete redirect-chassis way to configure gateways.
OVN has three ways to configure distributed router ports:
redirect-chassis, Gateway_Chassis, and HA_Chassis_Group.  They all end
up doing the same thing, so it's a lot of redundancy that exists only
for historical reasons.  This commit removes the oldest and least
general method, redirect-chassis, and converts all of the documentation
and tests to use newer methods.

This prepares for the ddlog implementation of ovn-northd, which doesn't
support redirect-chassis.  It *could* support redirect-chassis, but it
seemed to make sense to clean this up beforehand.

Signed-off-by: Ben Pfaff <blp@ovn.org>
Acked-by: Han Zhou <hzhou@ovn.org>
Acked-by: Numan Siddique <numans@ovn.org>