system tests: Fix occasional failure of the test - "Load balancer health checks".
Due to some timing issues, ovn-controller logs the warning message -
"handle service check: Service monitor not found". This can happen
if the service monitor is created in SB DB before the port for
the service is bound. This patch adds this warning message to the
WHITELIST of OVS_TRAFFIC_VSWITCHD_STOP.

Signed-off-by: Numan Siddique <numans@ovn.org>
Acked-by: Dumitru Ceara <dceara@redhat.com>