northd: Enhance implementation of datapath tunnel key requests.
When a tunnel key was requested, the implementation was not smart enough
to reassign a datapath that had been automatically assigned the same
key.  This fixes the problem and adds a test.

I didn't see a reason not to make this feature available for logical
routers so I added that feature too.

Signed-off-by: Ben Pfaff <blp@ovn.org>
Acked-by: Numan Siddique <numans@ovn.org>