Add SCTP load balancer test.
This is essentially a copy of existing load balancer tests that use TCP,
but with a few modifications for SCTP.

Userspace conntrack and SCTP do not currently mix. Therefore, the test
has been added to a new kernel-only system testsuite file. If userspace
conntrack support for SCTP is added in the future, we can potentially
move the SCTP test into the unified system-ovn.at file.

Signed-off-by: Mark Michelson <mmichels@redhat.com>
Acked-by: Numan Siddique <numans@ovn.org>
Signed-off-by: Numan Siddique <numans@ovn.org>