ovn-northd: Optimize flows for LB Hairpin traffic.
In order to detect that traffic was hairpinned due to logical switch load
balancers we need to match on source and destination IPs of packets (and
protocol ports potentially) in table ls_in_pre_hairpin.

For this, until now, we created 2 logical flows for each backend of a load
balancer VIP. However, in scenarios where large load balancers (i.e.,
with large numbers of backends) are applied to multiple logical
switches, this might generate logical flow count explosion.

One optimization is to generate a single logical flow per VIP that
combines all conditions generated for each backend. This reduces load on
the SB DB because of lower number of logical flows and also reduces
overall DB size because of less overhead due to other fields on the
logical_flow records.

Comparison of various performance aspects when running OVN with the NB
database attached to the bug report on a deployment with all VIFs bound
to a single node (62 load balancer VIPs with 513 load balancer
backends, applied on 106 logical switches):

Without this patch:
- SB database size: 60MB
- # of pre-hairpin logical flows: 109074
- # of logical flows: 159414
- ovn-controller max loop iteration time when processing SB DB: 8803ms
- ovn-northd max loop iteration time: 3988ms

With this patch:
- SB database size: 29MB (~50% decrease)
- # of pre-hairpin logical flows: 13250 (~88% decrease)
- # of logical flows: 63590 (~60% decrease)
- ovn-controller max loop iteration time when processing SB DB: 5585ms
- ovn-northd max loop iteration time: 1594ms

Reported-by: Aniket Bhat <anbhat@redhat.com>
Reported-at: https://bugzilla.redhat.com/1827403
Fixes: 1be8ac65bc60 ("ovn-northd: Support hairpinning for logical switch load balancing.")
Signed-off-by: Dumitru Ceara <dceara@redhat.com>
Signed-off-by: Numan Siddique <numans@ovn.org>