ofctrl: Split large group_mod messages up.
Group mod messages have the possibility of growing very large if OVN
installs a load balancer with a great many backends. The current
approach is to send a single ADD message with the entire group contents.
If the size of this message exceeds UINT16_MAX, then OpenFlow cannot
properly express the length of the message since the OpenFlow header's
length is limited to 16 bits.

This patch solves the problem by breaking the message into pieces. The
first piece is an ADD, and subsequent messages are INSERT_BUCKET
messages. This way, we end up being able to express the entire size of
the group through multiple OpenFlow messages.

Signed-off-by: Mark Michelson <mmichels@redhat.com>
Acked-by: Numan Siddique <numans@ovn.org>