Add info to help understand VTEP gateway easier
It is a must to set the column of the "addresses" to "unknown".
It will add a priority 0 entry in "ls_in_l2_lkup" stage.
If not, the  traffic can't go to physical network. I think
it is better to explain in this doc, so the player of OVN
can be easier to use VTEP gateway.

Author: Tao YunXiang <taoyunxiang@cmss.chinamobile.com>
Signed-off-by: Tao YunXiang <taoyunxiang@cmss.chinamobile.com>
Signed-off-by: Numan Siddique <numans@ovn.org>