ovn-northd.at: Fix race condition in pause and resume test.
A recent build failed because the pause and resume test failed. After
resuming northd and northd-backup, the status of northd was "standby"
instead of the expected "active". This happened for one of two reasons:

1) For some reason, northd failed to retrieve the lock between resuming
   and checking its status, so it reported as being "standby".
2) northd-backup won the race to get the lock, so it became the active
   instance instead of northd.

This change alters the test to fix both of these issues. Instead of
resuming both northd and northd-backup, we start by only resuming
northd. We then wait until it reports that it is active before we resume
northd-backup. This way, they are guaranteed to be in the states we
expect.

Signed-off-by: Mark Michelson <mmichels@redhat.com>
Acked-by: Numan Siddique <numans@ovn.org>