Add northd and ovn-controller cluster status reset commands.
During the course of debugging a clustered DB environment, all members
of the southbound database cluster were destroyed (i.e. the .db files
were removed from disk) and then restarted. Once this happened,
ovn-northd and ovn-controller could not interact with the southbound
database because they both detected all members of the cluster as having
"stale" data. The only course of action was to reset ovn-northd and all
ovn-controllers. It is possible to have this happen with the northbound
database as well if it is clustered.

This patch offers new ovn-appctl commands for ovn-northd and
ovn-controller that allows for it to reset its clustered status. This
allows for it to interact with the database successfully after a cluster
teardown and restart.

Signed-off-by: Mark Michelson <mmichels@redhat.com>
Acked-by: Han Zhou <hzhou@ovn.org>
Reported-at: https://bugzilla.redhat.com/show_bug.cgi?id=1829109