Fix the test case "80. ovn -- 4 HV, 1 LS, 1 LR, packet test with HA distributed router gateway port"
This test case is failing intermittently in my local setup.

***********************
rcv_n=3 exp_n=2
ovn.at:12: wait succeeded immediately
../../tests/ovn.at:9473: sort $rcv_text
--- expout	2020-06-17 10:57:23.467337249 +0530
+++ tests/testsuite.dir/at-groups/80/stdout	2020-06-17 10:57:23.469337182 +0530
@@ -1,2 +1,3 @@
 f0000001020400000201020308004500001c000000003f110100c0a80102ac1001030035111100080000
 ffffffffffff00000201020308060001080006040001000002010203ac100101000000000000ac100101
+ffffffffffff00000201020308060001080006040001000002010203ac100101000000000000ac100101
*******************

Before calling 'OVN_CHECK_PACKETS', there is a sleep of 1 second. This is not required,
and looks like this sleep is causing an extra ARP packet to be received.

This patch deletes this sleep. The macro OVN_CHECK_PACKETS will anyway wait for the expected
packets to be received.

Acked-by: Mark Michelson <mmichels@redhat.com>
Signed-off-by: Numan Siddique <numans@ovn.org>