Support packet metadata marking for logical router policies.
This patch adds a new column 'options' of type smap in the
Logical_Router_Policy table in the NB DB and supports the key 'pkt_mark'.
CMS can set a desired value for this key in the 'options' column. When this
router policy is applied, the packet metadata is marked with the specified
value (to the NXM_NX_PKT_MARK OVS field).

In the case of Linux, this corresponds to struct sk_buff's "skb_mark"
member and this mark can be seen by the linux networking subsystem.
CMS can inspect this value (as an iptables rule or adding an OF flow
in another ovs bridge) and take appropriate action when the marked packet
leaves the integration bridge via the patch port.

Requested-at: https://bugzilla.redhat.com/show_bug.cgi?id=1828933
Requested-by: Alexander Constantinescu <aconstan@redhat.com>
Acked-by: Dumitru Ceara <dceara@redhat.com>
Signed-off-by: Numan Siddique <numans@ovn.org>