ovn-northd: Fix the missing lflow issue in LS_OUT_PRE_LB.
When load balancer(s) configured with VIPs are associated to a logical switch,
then ovn-northd adds the below logical flow so that the packets in the egress
switch pipeline enter the conntrack.

table=0 (ls_out_pre_lb      ), priority=100  , match=(ip), action=(reg0[[0]] = 1; next;)

ovn-northd maintains a local boolean variable 'vip_configured' in
build_pre_lb() and adds the above lflow if this is true at the end.
But this variable is overriden as -> vip_configured = !!lb->n_vips;
when it loops through every load balancer associated with the logical switch.

This is wrong and this patch fixes this issue.

A test case is addd to test this scenario and the test case fails without the
fix in this patch.

Fixes: bb9f2b9ce56c("ovn-northd: Consider load balancer active backends in router pipeline")

Reported-at: https://bugzilla.redhat.com/show_bug.cgi?id=1849162
Reported-by: Tim Rozet <trozet@redhat.com>
Acked-by: Dumitru Ceara <dceara@redhat.com>
Signed-off-by: Numan Siddique <numans@ovn.org>