ovn-controller: Fix the missing flows with monitor-all set to True
When the config ovn-monitor-all is set to True, then ovn-controller is not
programming the flows completely when it binds the logical port of
a logical switch until a full recompute is triggered.

This issue was introduced by the commit in Fixes - which added incremental procesing
for runtime data changes.

This patch fixes this issue.

Fixes: 3d096f833c4a("ovn-controller: Handle runtime data changes in flow output engine")

Reported-by: Dumitru Ceara <dceara@redhat.com>
Acked-by: Dumitru Ceara <dceara@redhat.com>
Signed-off-by: Numan Siddique <numans@ovn.org>

(cherry-picked from master commit f8a81693b05891fe0717ee4957b651f71dd964f5)