chassis: Propagate ovn-monitor-all external-id to Chassis:other_config.
At least for avoiding races in the unit tests it's useful to store the value
of the OVS ovn-monitor-all external-id in the Chassis record other_config
field. This allows us to know for sure when an ovn-controller has processed
the update to the OVS Open_vSwitch DB table.

Signed-off-by: Dumitru Ceara <dceara@redhat.com>
Signed-off-by: Numan Siddique <numans@ovn.org>

(cherry-picked from master commit 2387079f54427bb03721fc85f245e85aaed3b624)