ovn-controller: Cache logical flow expr matches.
This patch is a second attempt in caching the expr tree of logical flows.
Earlier attemp [1] was reverted.

In this approach, we do the following

  - If a logical flow match has any port group/address set references, we don't
    cache expr match or expr tree.

  - If a logical flow match doesn't have a port group/address set reference and
    expr condition match - is_chassis_resident, we cache the expr matches
    (which is hmap of struct expr_match *).

  - If a logical flow match has expr condition match - is_chassis_resident we
    cache the simplified 'expr' tree.

Caching is configurable and can be disabled if desired. The second patch of this series
added this support.

In my testing, a complete lflow_run() took around 5 seconds for around 90k logical flows,
where as it took around 2 seconds with caching.

Acked-by: Mark Michelson <mmichels@redhat.com>
Signed-off-by: Numan Siddique <numans@ovn.org>