lflow.c: No need to remove flows for adding new datapath.
When handling a new datapath, the flows should have never existed, so there is
no need to remove them before adding. Although it seems not harmful to do it,
the implementation was not complete. To remove existing flows, it also need to
remove the flow references. Because the flows never existed, so this wasn't a
problem. So, instead of fixing the incomplete flow removing, this patch simply
avoid the unnecessary operation.

Acked-by: Mark Michelson <mmichels@redhat.com>
Signed-off-by: Han Zhou <hzhou@ovn.org>