ovn-controller: Fix conjunction handling with incremental processing.
When translating lflows to OVS flows, different lflows can refer to same OVS
flow as a result of calling ofctrl_add_or_append_flow(), particularly for
conjunction combinding. However, the implementation doesn't work with
incremental processing, because when any of the lflows are removed, we rely on
the lflow's uuid to remove the OVS flow in the desired flow table. Currently
only one single lflow uuid is maintained in the desired flow, so removing one
of the lflows that references to the same desired flow resulted in wrong
behavior: either removing flows that are used by other lflows, or the existing
flows are not updated (part of the conjunction actions should be removed from
the flow).

To solve the problem, this patch maintains the cross reference (M:N) between
lflows (and other sb objects) and desired flows, and handles the flow removal
and updates with a flood-removal and re-add approach.

Fixes: e659bab31a9 ("Combine conjunctions with identical matches into one flow.")
Cc: Mark Michelson <mmichels@redhat.com>
Acked-by: Mark Michelson <mmichels@redhat.com>
Signed-off-by: Han Zhou <hzhou@ovn.org>