ovn-controller: Disable ofctrl probe by default.
The commit c99069c8934c ("ovn-controller: Provide the option to configure
inactivity probe interval for OpenFlow conn") supports configuring probe
interval for the openflow connection of ofctrl module. However, it changed the
default behavior. The default OpenFlow connection probe was defaults to 0
(disabled), because the connection is always unix socket (to local OVS br-int
bridge).  After that patch it changed to probe every 5 seconds, which was not
the intension.

This patch restores the default behavior of disabling the probe by setting
default value to 0, while keeping the capability of letting user configure it.

Cc: Numan Siddique <nusiddiq@redhat.com>
Acked-by: Mark Michelson <mmichels@redhat.com>
Signed-off-by: Han Zhou <hzhou@ovn.org>