ovn-northd: Replace ds_put_* evaluating to constant expressions
Replaces ds_put_cstr and ds_put_format which evaluate to constant
expressions with string constants

This version also removes invocations of ds_clear() which are no
longer required as per comments by I. Maximets on the first version.

It also covers one more case which was missed in the first version.

Signed-off-by: Anton Ivanov <anton.ivanov@cambridgegreys.com>
Acked-by: Numan Siddique <numans@ovn.org>
Signed-off-by: Mark Michelson <mmichels@redhat.com>