actions: Add a new OVN action - reject {}.
This action is similar to tcp_reset and icmp4/icmp6 action. If the matching
packet is an IPv4 of IPv6 TCP packet, it sends out TCP RST packet else it
sends out ICMPv4 or ICMPv6 Destination unreachable packet.

While transforming the original packet to the reject packet, this action
implicitly swaps the eth source with eth destination and IP source with
IP destination.

A future patch will make use of this action for reject ACL.

Acked-by: Mark Michelson <mmichels@redhat.com>
Acked-by: Dumitru Ceara <dceara@redhat.com>
Signed-off-by: Numan Siddique <numans@ovn.org>