Fix OVN update issue when ovn-controller is updated first from 20.06 to 20.09.
The commit in the Fixes tag changed the ct_commit signature from
ct_commit(..) to ct_commit{ ..}.  When ovn-controllers are updated
to a vesion which has this change and if ovn-northd is still not
udated to this version, then the logical flow with the action
ct_commit(..) will be rejected by ovn-controllers resulting in
datapath disruptions.  OVN recommends ovn-controller updates before
ovn-northd, but it is broken now.  This patch fixes this issue by
adding back the support for the old ct_commit(..).  We now support
both the versions of ct_commit.

Fixes: 6cfb44a76c61("Used nested actions in ct_commit)
CC: Mark Michelson <mmichels@redhat.com>
Signed-off-by: Numan Siddique <numans@ovn.org>
Acked-by: Flavio Fernandes <flavio@flaviof.com>
Signed-off-by: Numan Siddique <numans@ovn.org>

(cherry-picked from master commit 4799f73fd6edbab01f33ffc181b12f24f55b2ac7)