vtep tests: Fix the failing test - ovn-controller-vtep - binding 1.
The below check in the test - "ovn-controller-vtep - binding 1." is
failing intermittently after the commit [1].

Before the commit[1], the check was -
  ovn-sbctl --timeout=10 wait-until Port_Binding br-vtep_lswitch0_dup chassis!='[[]]'

And after this commit, the check is -
  check ovn-sbctl wait-until Port_Binding br-vtep_lswitch0_dup chassis!='[[]]'

The above check expects 'br-vtep_lswitch0_dup' to be
claimed by the chassis, but there is already a logical
port - 'br-vtep_lswitch0' which is claimed by the same
chassis.  Both these logical ports are set with the same
options, which means out of lports - 'br-vtep_lswitch0' and
'br-vtep_lswitch0_dup', only one will be claimed.

Looking into the code, the lport to be claimed is not
deterministic.  And that's why the test succeeds sometimes.
There are 2 possibilities
  - ovn-controller-vtep releases 'br-vtep_lswitch0' and claims
'br-vtep_lswitch0_dup' or
  - ovn-controller-vtep doesn't claim 'br-vtep_lswitch0_dup' at all
    (and doesn't release 'br-vtep_lswitch0').

This patch removes the above check as we can't be sure that
'br-vtep_lswitch0_dup' will be claimed all the time.  The
test will now ensure that only one is claimed.

[1] - 5fedcce6c3e0("tests: Remove explicit timeouts from most utility invocations.")
Signed-off-by: Numan Siddique <numans@ovn.org>
Acked-by: Dumitru Ceara <dceara@redhat.com>