controller: Add support for Logical Datapath Groups.
ovn-controller will receive updates from Logical_DP_Group table
and process logical flows accordingly.  Feature is fully backward
compatible since old 'logical_datapath' column kept as is.
It will also be used by nothd to not create datapath groups if there
is ony one datapath in it.

Unfortunately, almost every part of the ovn-controller depends on
fact that there is 1:1 relation between logical flows and logical
datapaths, starting from the logical flow handling and all the way
to deep internals of expression parsing and I-P engine.
So, instead of re-writing everything we're taking a "safe" approach
and just re-factoring a bit to add new 'datapath' arguments to
functions and call them in a loop for all datapaths in a datapath
group.  This might have some performance impact in case datapath
groups are actually used by nothd and there are many datapaths that
are local to this ovn-controller.  However, this imapct might be
compensated by lower number of logical flows in general.
There should be no performance penalty if datapath groups are not
used by northd.

Signed-off-by: Ilya Maximets <i.maximets@ovn.org>
Signed-off-by: Mark Michelson <mmichels@redhat.com>
Acked-by: Dumitru Ceara <dceara@redhat.com>