lflow: Relax the condition that detects Load_Balancer hairpin.
Commit fc219d84b667 [0] optimized Load Balancer hairpin support from an
OVN Southbound perspective by moving generation of flows to detect
hairpin traffic in ovn-controller.

However, this can be further optimized by relaxing the condition for
flows that detect hairpin traffic in the "original" direction.

It is safe to exclude the match on datapath.tunnel_key in the original
direction if we make sure that we only match on traffic that was
successfully load balanced (i.e., ct.natted = 1).  At this point it
would be enough to check that ip.src == ip.dst in order to determine
that traffic needs to be hairpinned.  Because OVS doesn't allow such
checks between fields we instead check against the known load balancer
backend IPs: ip.src == backend-ip && ip.dst == backend-ip.

This change improves scalability by reducing the number of hairpin OF
flows by ~50%.  For example, on a setup with:
- N Load Balancer VIPs.
- M Backends per VIP.
- All N VIPs applied to S logical switches.

Before this change the number of hairpin OF flows was:
- N x M x S (for original direction) + N x M x S (for reply direction)

With this change the number of hairpin OF flows is:
- N x M (for original direction ) + N x M x S (for reply direction)

That means that on a setup with N=100 VIPs, M=10 backends, S=100
switches the number of hairpin OF flows will be 101K vs 200K before this
change.

[0] fc219d84b667 ("actions: Add new actions chk_lb_hairpin, chk_lb_hairpin_reply and ct_snat_to_vip.")

Signed-off-by: Dumitru Ceara <dceara@redhat.com>
Signed-off-by: Numan Siddique <numans@ovn.org>

(cherry-picked from master commit c3d6c718856f43a35f3c8972e9cbe17b6606798e)