nbctl: Remove column verification for partial updates.
Since this is not a read-modify-write sequence, it's not strictly
necessary to verify that current value is exactly same while performing
mutations.  Verification removal will allow to significantly reduce
size of ovsdb transaction, because it will no longer contain all the
data from the mutated column.

Before this change, addition of 1 new switch port to a logical switch
with 1000 ports looked like this:

  {transact,
      where: _uuid  == 'logical switch row uuid'
      op   : wait
      until: ==
      rows : ports ['< list of current 1000 ports >']

      < create new lsp in Logical_Switch_Port table >

      where: _uuid  == 'logical switch row uuid'
      op   :  mutate
      mutations : ports insert ['< 1 uuid of a new lsp >']
  }

After:

  {transact,
      < create new lsp in Logical_Switch_Port table >

      where: _uuid  == 'logical switch row uuid'
      op   :  mutate
      mutations : ports insert ['< 1 uuid of a new lsp >']
  }

This should relieve some pressure from the ovsdb-server since it will
not need to parse and execute this huge 'wait' operation.

Signed-off-by: Ilya Maximets <i.maximets@ovn.org>
Signed-off-by: Numan Siddique <numans@ovn.org>