ovn-controller.at: Fix occasionally failing unit test.
Test "ovn-controller - Chassis other_config" occasionally fails because
it assumes the order of Encaps is always "geneve" first.  There is
however no guarantee for the iteration order of OVSDB records and in
some cases "vxlan" is returned as first Encap generating a slightly
different error message.

The purpose of the test is to detect the constraint violation so we can
relax the grep match.

Fixes: fc359bfe9342 ("chassis: Do not try to guess system-id changes.")
Signed-off-by: Dumitru Ceara <dceara@redhat.com>
Acked-by: Mark Michelson <mmichels@redhat.com>
Signed-off-by: Numan Siddique <numans@ovn.org>