Implement SCTP-specific reject() action.
Currently in OVN, if an SCTP packet hits a reject() action, OVN responds
with an ICMP packet. Instead, we should send an SCTP packet with an
ABORT chunk. This will either end the current association or will
prevent an association from being created, depending on which stage of
the SCTP state machine we currently are in.

This patch adds the desired behavior for SCTP. The reject() action will
now send an SCTP ABORT if the incoming packet is SCTP.

Signed-off-by: Mark Michelson <mmichels@redhat.com>
Acked-by: Numan Siddique <numans@ovn.org>